Draw line (and area) charts as step functions
These charts are used for things like the balance in an account, and the
data points are at changes in the balance.  The line from each data
point should then extend right horizontally because the balance does not
change until the next data point.

There may be some accounts where interpolation is appropriate (e.g.
salary income under accrual accounting, where the postings are samples
and should be interpolated linearly) but this is not the general case.
Also Beancount does no interpolation, so the stepwise line more
accurately represents Beancount's view of the account.